Move test support code to tests/lib
